# perf(rust, python): speed up from_par_iter Option<bool> `~2.5x`