# master

```
null density: 0 took 828ms
null density: 0.2 took 960ms
null density: 0.5 took 1035ms
null density: 0.8 took 660ms
```

# PR

```
null density: 0 took 308ms
null density: 0.2 took 406ms
null density: 0.5 took 405ms
null density: 0.8 took 307ms
```